+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
  -->
 
+## [1.0.7] - 2023-11-11
+
+### Changed
+
+- removed minify from rollup config
+
 ## [1.0.6] - 2023-11-11
 
 ### Fixed